I DID IT
Got the text file translator working and now I am able to get the info of the pokemon by merely entering the name of the pokemon and the stat I want from it. Through encapsulation the interface of the methods is very clean making it very readable and very minimal.